Refactor dealing with instances and repositories
Previously, all of that was done in the Repository class (or its
subclasses). We got away with it by letting instance configuration
“trickle down” into the repositories, but further processed things only
on the repository level. Making instances objects in their own right
enables querying forges for repositories beneath an organization
(GitHub) or namespace (Pagure) and create repository objects out of the
results, which can then be iterated over along with those explicitly
defined in configuration.

Along with that, it stood out that labelling JIRA tickets only by their
repository name/path can be ambiguous when dealing with more than one
instance, therefore JIRA tickets are labelled by instance and repository
and queried by that plus only the repository as a fallback now.

Also, include GitHub-related code paths when testing configuration and
the main synchronization function.

Signed-off-by: Nils Philippsen <nils@redhat.com>