Don’t ignore API errors in get_open_issues()
Previously, e.g. a 404 “Not Found” error would throw
get_next_page()-based pagination into endless loops, and it would not
have worked around connection timeouts anyway.

Instead, raise an exception on status codes of the 400 and 500 range,
except on 404 on the first page: This is because Pagure repositories
without issues enabled don’t report this in the repository metadata,
i.e. with querying repositories in bulk, if there is one without issues,
the only way to deal with it is to report back “no issues”.

Also, make looping over pages a bit simpler.

Signed-off-by: Nils Philippsen <nils@redhat.com>