Fix the issue with tickets not transitioning to new state correctly
The issue was with how the jira statuses configuration was handled after the
refactoring. It didn't returned list of states anymore, but rather just a string
which didn't matched against the status in JIRA.

I also updated the .gitignore with new config files.

Removing the verbose mode from the github action as it was used just to
troubleshoot this issue.

Moved few log messages from debug to info as they will be crucial in identifying
the issue.

Signed-off-by: Michal Konecny <mkonecny@redhat.com>