Add implementation of story_points
This commit adds implementation of story points to jira. The jira_sync
script will now look for specific labels in github and pagure and
convert them to story points value. This is configurable.

The field which is used in JIRA is configurable as well, if left empty
it will be skipped completelly.

Closes #108

Signed-off-by: Michal Konecny <mkonecny@redhat.com>