Add a Fedora Messaging consumer to sync Bugzilla tickets
Bugzilla tickets should be created in Jira if they have the `mirror`
flag set to `+`. Created Jira tickets should have the Bugzilla URL in
the summary field. Then Watson will sync the ticket automatically.

This commit adds a Fedora Messaging consumer that listens to Bugzilla
events where the `mirror` flag is set to `+` and creates and issue in
Jira with the correct `summary`.

Signed-off-by: Aurélien Bompard <aurelien@bompard.org>