Prevent creating JIRA issue multiple times
Recently there was an outage of sync from upstream automation on JIRA
and that caused some tickets to be created over and over again till I
manually filled the External Issue URL field. This change should prevent
that by checking if the url is in summary in case it's not in External
Issue URL field.

Signed-off-by: Michal Konecny <mkonecny@redhat.com>